Implemented debugger modules and rename files
