+++ src/http/common/Debugger.ts
@@ -1,0 +1,9 @@
+/**
+ * ! Module debugger to show logs in cosole
+ * * whitehatdevv - 2021/12/13
+ */
+export module Debuuger {
+
+    
+
+}
